+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Fixed
 
 ### Changed
+* Move MCP server to `neocarta` library
+* Change MCP server name to `neocarta-mcp`
+* Update MCP server imports in `eval/` module
 
 ### Added